+++ src/Bridge/AzureSearch/SearchStore.php
@@ -16,7 +16,6 @@ use Symfony\AI\Platform\Vector\Vector;
 use Symfony\AI\Store\Document\Metadata;
 use Symfony\AI\Store\Document\VectorDocument;
 use Symfony\AI\Store\StoreInterface;
-use Symfony\Component\Uid\Uuid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
@@ -95,7 +94,7 @@ final class SearchStore implements StoreInterface
     private function convertToVectorDocument(array $data): VectorDocument
     {
         return new VectorDocument(
-            id: Uuid::fromString($data['id']),
+            id: $data['id'],
             vector: !\array_key_exists($this->vectorFieldName, $data) || null === $data[$this->vectorFieldName]
                 ? new NullVector()
                 : new Vector($data[$this->vectorFieldName]),