+++ src/Bridge/Milvus/Store.php
@@ -18,7 +18,6 @@ use Symfony\AI\Store\Document\VectorDocument;
 use Symfony\AI\Store\Exception\InvalidArgumentException;
 use Symfony\AI\Store\ManagedStoreInterface;
 use Symfony\AI\Store\StoreInterface;
-use Symfony\Component\Uid\Uuid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
@@ -155,7 +154,7 @@ final class Store implements ManagedStoreInterface, StoreInterface
     private function convertToIndexableArray(VectorDocument $document): array
     {
         return [
-            'id' => $document->id->toRfc4122(),
+            'id' => $document->id,
             '_metadata' => json_encode($document->metadata->getArrayCopy()),
             $this->vectorFieldName => $document->vector->getData(),
         ];
@@ -174,6 +173,6 @@ final class Store implements ManagedStoreInterface, StoreInterface
 
         $score = $data['distance'] ?? null;
 
-        return new VectorDocument(Uuid::fromString($id), $vector, new Metadata(json_decode($data['_metadata'], true)), $score);
+        return new VectorDocument($id, $vector, new Metadata(json_decode($data['_metadata'], true)), $score);
     }
 }