+++ src/Bridge/Typesense/Store.php
@@ -18,7 +18,6 @@ use Symfony\AI\Store\Document\VectorDocument;
 use Symfony\AI\Store\Exception\InvalidArgumentException;
 use Symfony\AI\Store\ManagedStoreInterface;
 use Symfony\AI\Store\StoreInterface;
-use Symfony\Component\Uid\Uuid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
@@ -119,7 +118,7 @@ final class Store implements ManagedStoreInterface, StoreInterface
     private function convertToIndexableArray(VectorDocument $document): array
     {
         return [
-            'id' => $document->id->toRfc4122(),
+            'id' => $document->id,
             $this->vectorFieldName => $document->vector->getData(),
             'metadata' => json_encode($document->metadata->getArrayCopy()),
         ];
@@ -140,6 +139,6 @@ final class Store implements ManagedStoreInterface, StoreInterface
 
         $score = $data['vector_distance'] ?? null;
 
-        return new VectorDocument(Uuid::fromString($id), $vector, new Metadata(json_decode($document['metadata'], true)), $score);
+        return new VectorDocument($id, $vector, new Metadata(json_decode($document['metadata'], true)), $score);
     }
 }