+++ src/Bridge/Weaviate/Store.php
@@ -18,7 +18,6 @@ use Symfony\AI\Store\Document\VectorDocument;
 use Symfony\AI\Store\Exception\InvalidArgumentException;
 use Symfony\AI\Store\ManagedStoreInterface;
 use Symfony\AI\Store\StoreInterface;
-use Symfony\Component\Uid\Uuid;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
@@ -120,10 +119,10 @@ final class Store implements ManagedStoreInterface, StoreInterface
     {
         return [
             'class' => $this->collection,
-            'id' => $document->id->toRfc4122(),
+            'id' => $document->id,
             'vector' => $document->vector->getData(),
             'properties' => [
-                'uuid' => $document->id->toRfc4122(),
+                'uuid' => $document->id,
                 'vector' => $document->vector->getData(),
                 '_metadata' => json_encode($document->metadata->getArrayCopy()),
             ],
@@ -141,6 +140,6 @@ final class Store implements ManagedStoreInterface, StoreInterface
             ? new NullVector()
             : new Vector($data['vector']);
 
-        return new VectorDocument(Uuid::fromString($id), $vector, new Metadata(json_decode($data['_metadata'], true)));
+        return new VectorDocument($id, $vector, new Metadata(json_decode($data['_metadata'], true)));
     }
 }