+++ src/Document/TextDocument.php
@@ -20,7 +20,7 @@ use Symfony\Component\Uid\Uuid;
 final class TextDocument implements EmbeddableDocumentInterface
 {
     public function __construct(
-        private readonly Uuid $id,
+        private readonly int|string|Uuid $id,
         private readonly string $content,
         private readonly Metadata $metadata = new Metadata(),
     ) {
@@ -34,7 +34,7 @@ final class TextDocument implements EmbeddableDocumentInterface
         return new self($this->id, $content, $this->metadata);
     }
 
-    public function getId(): Uuid
+    public function getId(): int|string|Uuid
     {
         return $this->id;
     }